+++ src/lib/core.ts
@@ -8,5 +8,6 @@ export * from "./core/axis";
 export * from "./core/constants";
 export * from "./core/pie";
 export * from "./core/primitives";
+export * from "./core/scales";
 export * from "./core/series";
 export * from "./core/timeSeries";